feat: Add ContactForm unit test
