+++ ajax/axiosMockAdapter.ts
@@ -2,10 +2,12 @@
 import MockAdapter from 'axios-mock-adapter';
 import axios from 'axios';
 
-export const initializeAxiosMockAdapter = (): void => {
+export const initializeAxiosMockAdapter = (delay = 0, isSuccess = true): void => {
     const mock = new MockAdapter(axios, {
-        delayResponse: 1000,
+        delayResponse: delay,
     });
 
-    mock.onPost('https://aguarela-contact-form-backend.herokuapp.com/api/sendEmail').reply(200, {});
+    const httpStatusCode = isSuccess ? 200 : 500;
+
+    mock.onPost('https://aguarela-contact-form-backend.herokuapp.com/api/sendEmail').reply(httpStatusCode, {});
 };
