+++ package.json
@@ -41,8 +41,10 @@
         "styled-jsx-plugin-sass": "^1.0.0"
     },
     "devDependencies": {
+        "@testing-library/dom": "^8.1.0",
         "@testing-library/jest-dom": "^5.14.1",
         "@testing-library/react": "^12.0.0",
+        "@testing-library/user-event": "^13.2.1",
         "@types/classnames": "^2.2.11",
         "@types/jest": "^26.0.24",
         "@types/node": "^14.14.25",
